aisi_utils/config_set/requirements.txt
azure-common==1.1.28
azure-core==1.23.0
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-identity==1.8.0
azure-keyvault==4.1.0
azure-keyvault-certificates==4.3.0
azure-keyvault-keys==4.4.0
azure-keyvault-secrets==4.3.0
azure-nspkg==3.0.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
idna==3.3
isodate==0.6.1
msal==1.17.0
msal-extensions==0.3.1
msrest==0.6.21
oauthlib==3.2.0
portalocker==2.4.0
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
typing-extensions==4.1.1
urllib3==1.26.8
